hasibe working code
