+++ Testing2/TestingCms.csproj
@@ -50,6 +50,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="CMS.cs" />
     <Compile Include="UtCms.cs" />
   </ItemGroup>
   <ItemGroup>
